+++ controllers/PermissionController.php
@@ -38,13 +38,17 @@ class PermissionController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new TbTemplatePermissionSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if (\Yii::$app->user->can('view-project-template')) {
+            $searchModel = new TbTemplatePermissionSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else{
+            return $this->render('../error');
+        }
     }
 
     /**
@@ -55,9 +59,13 @@ class PermissionController extends Controller
      */
     public function actionView($template_id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($template_id),
-        ]);
+        if (\Yii::$app->user->can('view-project-template')) {
+            return $this->render('view', [
+                'model' => $this->findModel($template_id),
+            ]);
+        } else{
+            return $this->render('../error');
+        }
     }
 
     /**
@@ -67,19 +75,23 @@ class PermissionController extends Controller
      */
     public function actionCreate()
     {
-        $model = new TbTemplatePermission();
+        if (\Yii::$app->user->can('create-project-template')) {
+            $model = new TbTemplatePermission();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'template_id' => $model->template_id]);
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'template_id' => $model->template_id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
-        }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        } else{
+            return $this->render('../error');
+        }
     }
 
     /**
@@ -91,15 +103,19 @@ class PermissionController extends Controller
      */
     public function actionUpdate($template_id)
     {
-        $model = $this->findModel($template_id);
+        if (\Yii::$app->user->can('edit-project-template')) {
+            $model = $this->findModel($template_id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'template_id' => $model->template_id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'template_id' => $model->template_id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        } else{
+            return $this->render('../error');  
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -111,9 +127,13 @@ class PermissionController extends Controller
      */
     public function actionDelete($template_id)
     {
-        $this->findModel($template_id)->delete();
+        if (\Yii::$app->user->can('delete-project-template')) {
+            $this->findModel($template_id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else{
+            return $this->render('../error');  
+        }
     }
 
     /**
